I updated the name of a few functions in FM to have better descriptions.
Commit has issues:
when allowsCustomizations == false, selected notes are not being cleared.
I'm refactoring to simplify the process of adding notes to fretboard models.
Specifically I want to remove the isInScale bool since I will not be using the chromatic fill option that I was using in the desktop version.