Test Pass: all looks like it's working properly.
What I did:
----------------

Fixed problem that was causing the app to crash when when adding a newCollection.

delegate method: fbvc.collectionWasSelected(...) had the tableViews.reloadData() in the wrong position within the function. I'm not sure why that is.

fixed some duplicate code in CollectionsTVC.editCollection.

Changed	segmentedControl title from basic Chords, to chords.

FBCollectionModel: removed all references to images.

deletedUIView extension file which creates images from views.

FBVC: delete FBVC.viewWillDissapear.

added call in FBVC.viewDidLayoutSubviews to check if allowsCustomziations = false.
If so, it calls addNotesAction.
This is to ensure a scale is added to the fretboard upon the initial launch.

added 3 sets of	root notes in the pickerView, because I think it helps when cycling through chords.